+++ internal/db/sqltest/tests/credential/vault/credential_vault_library_username_password_mapping_override.sql
@@ -2,6 +2,8 @@
 --   the following triggers
 --    insert_credential_vault_library_mapping_override_subtype
 --    delete_credential_vault_library_mapping_override_subtype
+--   and the following view
+--    credential_vault_list_lookup_library
 
 begin;
   select plan(11);
@@ -16,14 +18,14 @@ begin;
     from credential_vault_library_mapping_override
    where library_id in ('vl______wvl4', 'vl______wvl5', 'vl______wvl6', 'vl______wvl7');
 
-  prepare select_private_libraries as
+  prepare select_libraries as
    select public_id::text, credential_type::text, username_attribute::text, password_attribute::text
-     from credential_vault_library_private
+     from credential_vault_list_lookup_library
     where public_id in ('vl______wvl2', 'vl______wvl3', 'vl______wvl4', 'vl______wvl5', 'vl______wvl6', 'vl______wvl7')
  order by public_id;
 
   select results_eq(
-      'select_private_libraries',
+      'select_libraries',
       $$VALUES
       ('vl______wvl2', 'unspecified',   null,          null),
       ('vl______wvl3', 'username_password', null,          null),